+++ R/use_yml.R
@@ -24,6 +24,9 @@
 #'   no template is set, checks `getOption("ymlthis.rmd_body")` (see
 #'   [`use_rmd_defaults()`]) and otherwise uses [`setup_chunk()`].
 #' @param quiet	Logical. Whether to message about what is happening.
+#' @param open_doc Logical. Open the document after it's created? By default,
+#'   this is `TRUE` if it is an interactive session and `FALSE` if not. Also
+#'   checks that RStudio is available.
 #'
 #' @return `use_yml()` invisibly returns the input `yml` object
 #' @export
@@ -35,7 +38,8 @@ use_yml <- function(.yml = last_yml()) {
 
 #' @rdname use_yml
 #' @export
-use_rmarkdown <- function(.yml = last_yml(), path, template = NULL, include_yaml = TRUE, include_body = TRUE, body = NULL, quiet = FALSE) {
+use_rmarkdown <- function(.yml = last_yml(), path, template = NULL, include_yaml = TRUE,
+                          include_body = TRUE, body = NULL, quiet = FALSE, open_doc = interactive()) {
 
   if (!is.null(template) && fs::is_dir(template)) {
     template_skeleton <- file.path(template, "skeleton", "skeleton.Rmd")
@@ -62,16 +66,26 @@ use_rmarkdown <- function(.yml = last_yml(), path, template = NULL, include_yaml
   }
 
   usethis::write_over(path, rmarkdown_txt, quiet = quiet)
-  if (rstudioapi::isAvailable() && interactive()) rstudioapi::navigateToFile(path, line = 2)
+  if (rstudioapi::isAvailable() && open_doc) rstudioapi::navigateToFile(path, line = 2)
 
   invisible(path)
 }
 
 #' @rdname use_yml
 #' @export
-use_index_rmd <- function(.yml = last_yml(), path = ".", template = NULL, quiet = FALSE) {
+use_index_rmd <- function(.yml = last_yml(), path, template = NULL, include_yaml = TRUE,
+                          include_body = TRUE, body = NULL, quiet = FALSE, open_doc = interactive()) {
   index_rmd_path <- file_path(path, "index.Rmd")
-  use_rmarkdown(.yml = .yml, path = index_rmd_path, template = template, quiet = quiet)
+  use_rmarkdown(
+    .yml = .yml,
+    path = index_rmd_path,
+    template = template,
+    include_yaml = include_yaml,
+    include_body = include_body,
+    body = body,
+    quiet = quiet,
+    open_doc = open_doc
+  )
 }
 
 combine_yml <- function(x, y) {
@@ -79,6 +93,7 @@ combine_yml <- function(x, y) {
 
   x
 }
+
 return_yml_code <- function(.yml) {
   yaml_text <- capture_yml(.yml)
   usethis::ui_code_block(yaml_text)
@@ -105,6 +120,12 @@ return_yml_code <- function(.yml) {
 #' [pkgdown](https://pkgdown.r-lib.org/articles/pkgdown.html) and
 #' [blogdown](https://bookdown.org/yihui/bookdown/configuration.html)
 #' documentation for more.
+#'
+#' By default, the yaml package adds a new line to the end of files. Some
+#' environments, such as RStudio Projects, allow you to append new lines
+#' automatically. Thus, you may end up with 2 new lines at the end of your file.
+#' If you'd like to automatically remove the last new line in the file, set
+#' `options(ymlthis.remove_blank_line = TRUE)`.
 #'
 #' @template describe_yml_param
 #' @param path a file path to write the file to
@@ -178,7 +199,8 @@ write_yml_file <- function(.yml, path, build_ignore = FALSE, git_ignore = FALSE,
 
   if (!is.null(.yml)) {
     if (is.character(.yml) && length(.yml) == 1) {
-      usethis::write_over(path, .yml, quiet = FALSE)
+      if (check_remove_blank_line()) .yml <- remove_blank_line(.yml)
+      usethis::write_over(path, .yml, quiet = quiet)
       return(invisible(path))
     }
 
@@ -188,7 +210,8 @@ write_yml_file <- function(.yml, path, build_ignore = FALSE, git_ignore = FALSE,
       column.major = FALSE
     )
 
-    usethis::write_over(path, yml_txt, quiet = FALSE)
+    if (check_remove_blank_line()) .yml <- remove_blank_line(.yml)
+    usethis::write_over(path, yml_txt, quiet = quiet)
     return(invisible(path))
   }
 
@@ -200,6 +223,21 @@ write_yml_file <- function(.yml, path, build_ignore = FALSE, git_ignore = FALSE,
 
 file_path <- function(path, .file) {
   file.path(normalizePath(path), .file)
+}
+
+check_remove_blank_line <- function() {
+  remove_line_opt <- getOption("ymlthis.remove_blank_line")
+  option_set <- !is.null(remove_line_opt)
+  if (option_set) {
+    if (!is.logical(remove_line_opt)) stop("option `ymlthis.remove_blank_line` must be either logical or `NULL`")
+    return(remove_line_opt)
+  }
+
+  FALSE
+}
+
+remove_blank_line <- function(x) {
+  stringr::str_remove(x, "\n$")
 }